Rename module element-connector-tcp into element-connector-tcp-netty.
