+++ demo-apps/cf-plugtest-client/pom.xml
@@ -22,7 +22,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
-			<artifactId>element-connector-tcp</artifactId>
+			<artifactId>element-connector-tcp-netty</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
